requirements.txt
pydantic>=2.0
openai>=1.30
structlog>=24.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
temporalio>=1.7
